// Client/src/app/api/model/meetup.ts
/**
 * Fitness Meetup
 * The public API for the FitnessMeetup platform
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';
import { User } from './user';


export interface Meetup { 
    id: number;
    title: string;
    description?: string;
    sport?: string;
    owner: User;
    participants?: Array<User>;
    date: Date;
    location: Location;
}